Check ConfirmEdit extension is loaded before getting its config value
The check to determine whether to use Captcha should not error out if
the "optional" ConfirmEdit extension is not installed. But at the moment
it does with fatal exception, because Config::get() will unconditionally
throw exception on attempt to get value with a key the object does not
know (which is the case with all globals of an extension not yet
installed).

Bug: T268966
Change-Id: I1044ac583ebd987dd3d3486931d037f260166276