Add config options to disable name, email, subject modification
On EnWiki, we found that allowing users to change their name and
email allows them to impersonate other users which in our usecase
would be a major problem.

This patch adds the following new config variables:
- NameReadonly: makes the Name field read-only (false by default)
- EmailReadonly: makes the Email field read-only (false by default)
- SubjectReadonly: makes the Subject field read-only (false by default)
- MustHaveEmail: whether to require the sender to have a confirmed
  email for accessing the form (false by default)

If the form allows anyone to submit, these config options do
nothing. If users are required to log in, and these options are set
to true, then these inputs are populated with the username and email
with editing of them disabled.

Bug: T324175
Change-Id: I4820c1effc213a2cba452f00b65c90d287ec2172