Start on selenium tests for ContactPage
Set up selenium testing on ContactPage and create the first test
that checks to see if the extension has been loaded and is visible
on Special:Version. This will be removed once other selenium tests
have been added, and is to ensure that the selenium tests are
working as expected.

Bug: T186864
Change-Id: I56cba33f9467957f407c1a441c9c3d7867cbe2a5