Define merge strategy for ContactConfig as array_plus_2d
Define a merge strategy for ContactConfig so that the config can
be changed easier. array_plus_2d is used so that the default can
be updated in a manner that allows setting the AdditionalFields
value to the empty array. Using array_replace_recursive would have
prevented removal of the "Text" additional field, whereas, array
_plus_2d bascially treats AdditionalFields as a scalar (as it's
array items are in the 3rd dimension).

Using this merge strategy is tested in phpunit tests that rely on
this patch.

Bug: T142546
Change-Id: Ic7e2bd99dc790df02831c7980a479742575b7cb5