Allow multiple messages to be form specific in the contact form
Allow many messages shown in Special:Contact to be form-specific.
This is done by using the method already used for messages like
contactpage-pagetext where if a message exists with with the name
of the form appended to the end (separated by a "-") it is used
instead of the generic one.

To achieve this a helper function is made that chooses the generic
or form-specific message following the rule above. If the generic
message is from core and therefore does not start with
"contactpage-" the form specific message is defined as
"contactpage-coremessagekey-formname".

Bug: T190553
Change-Id: Ifebfe64097a8f71f63f5c67d7097e0121cecd3d4