Treat temporary users as logged out for MustBeLoggedIn config
Treat temporary user accounts as logged out for the purposes of
the MustBeLoggedIn configuration for contact forms. This means
updating the call to ::requireLogin to instead call ::requireNamed
User. While doing this, the autofill of information (such as
email address and real name) is changed to only occur for named
users as temporary accounts do not have preferences and thus cannot
have an email and/or a defined real name.

Bug: T344722
Change-Id: I1ed557d00d7ee025b3c07d552e78ed46fd45d7aa