Add optional parameter to show message if user is blocked.
If a user or IP address is blocked from sending emails, they will not be
able to access the Contact form. For our purposes of account vanishing,
this is problematic because the user will not see any further
instructions on how to proceed with account vanishing, and would instead
see instructions for appealing the block.

This patch provides a simple mechanism for showing a message if the
account or IP is blocked (instead of the default block template), which
may contain other instructions for the user in specific cases like ours.

Bug: T343536
Change-Id: I77b4b049b07f9f1e729a9a0ca6b63a6d15343848