Add optional redirect parameter.
There are cases when we want a Contact form to become a redirect to another page. This can be to accommodate certain clients in the wild, such as older versions of native apps that can't be easily updated (or external websites with outdated links that we can't control), which navigate to a Contact form that we want to decommission.

Since Contact forms are not regular wiki pages that we can redirect with wikitext, we need to make an explicit configuration option for redirecting a Contact form to another arbitrary URL.

Bug: T372828
Change-Id: I34dcf18c7310e52c7814b3bed205547fe28dd4e7