Allow specifying a recipient email address
Previously, the extension only allowed you to specify a MediaWiki user account with a verified email address to receive the contact form messages. This commit provides the ability to specify an email address directly, bypassing the need for "dummy" wiki accounts that have typically been used in the past.

The name provided to the emailccsubject message is a new RecipientName config variable, which falls back to $wgSitename if one is not provided.

Bug: T201764
Change-Id: Ib384ef14dbbc9418c8415e862dd70a4ce79c4c0a